feat(chat): refactor chat ui code
- Implemented chat API endpoint for handling chat requests.
- Added Pydantic models for Vercel AI SDK UI types to facilitate message handling.
- Enhanced streaming response handling for chat interactions.
